requirements.txt
numpy
scikit-learn
sentence-transformers
faiss-cpu
streamlit
pytest
transformers
torch
